backend/requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
joblib==0.14.0
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
sklearn==0.0
sqlparse==0.3.0
